Create mesoscale utility script
Refs: #1
