+++ ush/mesoscale/mesoscale_util.py
@@ -1,0 +1,154 @@
+'''
+Name: mesoscale_util.py
+Contact(s): Mallory Row
+Abstract: Utilities for mesoscale component 
+'''
+
+import os
+import datetime
+
+def check_file(file_path):
+    """! Check file exists and not zero size
+         Args:
+             file_path - full path to file (string)
+         Returns:
+             file_good - full call to METplus (boolean)
+    """
+    if os.path.exists(file_path):
+        if os.path.getsize(file_path) > 0:
+            file_good = True
+        else:
+            file_good = False
+    else:
+        file_good = False
+    return file_good
+
+def format_filler(unfilled_file_format, valid_time_dt, init_time_dt,
+                  forecast_hour, str_sub_dict):
+    """! Creates a filled file path from a format
+         Args:
+             unfilled_file_format - file naming convention (string)
+             valid_time_dt        - valid time (datetime)
+             init_time_dt         - initialization time (datetime)
+             forecast_hour        - forecast hour (string)
+             str_sub_dict         - other strings to substitue (dictionary)
+         Returns:
+             filled_file_format - file_format filled in with verifying
+                                  time information (string)
+    """
+    filled_file_format = '/'
+    format_opt_list = ['lead', 'lead_shift', 'valid', 'valid_shift',
+                       'init', 'init_shift', 'cycle']
+    if len(list(str_sub_dict.keys())) != 0:
+        format_opt_list = format_opt_list+list(str_sub_dict.keys())
+    for filled_file_format_chunk in unfilled_file_format.split('/'):
+        for format_opt in format_opt_list:
+            nformat_opt = (
+                filled_file_format_chunk.count('{'+format_opt+'?fmt=')
+            )
+            if nformat_opt > 0:
+               format_opt_count = 1
+               while format_opt_count <= nformat_opt:
+                   if format_opt in ['lead_shift', 'valid_shift',
+                                     'init_shift']:
+                       shift = (filled_file_format_chunk \
+                                .partition('shift=')[2] \
+                                .partition('}')[0])
+                       format_opt_count_fmt = (
+                           filled_file_format_chunk \
+                           .partition('{'+format_opt+'?fmt=')[2] \
+                           .rpartition('?')[0]
+                       )
+                   else:
+                       format_opt_count_fmt = (
+                           filled_file_format_chunk \
+                           .partition('{'+format_opt+'?fmt=')[2] \
+                           .partition('}')[0]
+                       )
+                   if format_opt == 'valid':
+                       replace_format_opt_count = valid_time_dt.strftime(
+                           format_opt_count_fmt
+                       )
+                   elif format_opt == 'lead':
+                       if format_opt_count_fmt == '%1H':
+                           if int(forecast_hour) < 10:
+                               replace_format_opt_count = forecast_hour[1]
+                           else:
+                               replace_format_opt_count = forecast_hour
+                       elif format_opt_count_fmt == '%2H':
+                           replace_format_opt_count = forecast_hour.zfill(2)
+                       elif format_opt_count_fmt == '%3H':
+                           replace_format_opt_count = forecast_hour.zfill(3)
+                       else:
+                           replace_format_opt_count = forecast_hour
+                   elif format_opt == 'init':
+                       replace_format_opt_count = init_time_dt.strftime(
+                           format_opt_count_fmt
+                       )
+                   elif format_opt == 'cycle':
+                       replace_format_opt_count = init_time_dt.strftime(
+                           format_opt_count_fmt
+                       ) 
+                   elif format_opt == 'lead_shift':
+                       shift = (filled_file_format_chunk.partition('shift=')[2]\
+                                .partition('}')[0])
+                       forecast_hour_shift = str(int(forecast_hour)
+                                                 + int(shift))
+                       if format_opt_count_fmt == '%1H':
+                           if int(forecast_hour_shift) < 10:
+                               replace_format_opt_count = (
+                                   forecast_hour_shift[1]
+                               )
+                           else:
+                               replace_format_opt_count = forecast_hour_shift
+                       elif format_opt_count_fmt == '%2H':
+                           replace_format_opt_count = (
+                               forecast_hour_shift.zfill(2)
+                           )
+                       elif format_opt_count_fmt == '%3H':
+                           replace_format_opt_count = (
+                               forecast_hour_shift.zfill(3)
+                           )
+                       else:
+                           replace_format_opt_count = forecast_hour_shift
+                   elif format_opt == 'init_shift':
+                       shift = (filled_file_format_chunk.partition('shift=')[2]\
+                                .partition('}')[0])
+                       init_shift_time_dt = (
+                           init_time_dt + datetime.timedelta(hours=int(shift))
+                       )
+                       replace_format_opt_count = init_shift_time_dt.strftime(
+                           format_opt_count_fmt
+                       )
+                   elif format_opt == 'valid_shift':
+                       shift = (filled_file_format_chunk.partition('shift=')[2]\
+                                .partition('}')[0])
+                       valid_shift_time_dt = (
+                           valid_time_dt + datetime.timedelta(hours=int(shift))
+                       )
+                       replace_format_opt_count = valid_shift_time_dt.strftime(
+                           format_opt_count_fmt
+                       )
+                   else:
+                       replace_format_opt_count = str_sub_dict[format_opt]
+                   if format_opt in ['lead_shift', 'valid_shift', 'init_shift']:
+                       filled_file_format_chunk = (
+                           filled_file_format_chunk.replace(
+                               '{'+format_opt+'?fmt='
+                               +format_opt_count_fmt
+                               +'?shift='+shift+'}',
+                               replace_format_opt_count
+                           )
+                       )
+                   else:
+                       filled_file_format_chunk = (
+                           filled_file_format_chunk.replace(
+                               '{'+format_opt+'?fmt='
+                               +format_opt_count_fmt+'}',
+                               replace_format_opt_count
+                           )
+                       )
+                   format_opt_count+=1
+        filled_file_format = os.path.join(filled_file_format,
+                                          filled_file_format_chunk)
+    return filled_file_format
